Improved igturtle.goTo().  Added igturtle.refuel().
igturtle.goTo can now accept a single table argument with
members x and y (and optionally z) to define the
destination.

New function igturtle.refuel() will refuel the turtle in one of
two ways.  By default, will go to the predefined _posFuel
location, pull fuel from below, consume it, and return to its
previous location.  If the "enderfuel" option is given with a
numerical value, it will assume it has an enderchest in that
slot, place it down, pull fuel from it, refuel, and pick it up in
that slot again.