Added finer version control to ig.require().
The API will now track its version numbers.  When a user calls the
ig.require() method to load an API, an optional version string can be
supplied.  This can be one of three forms:
"major", "major.minor", or "major.minor.patch"
If the minor or patch numbers are omitted, the latest versions are
substituted.  For example. if the API has versions 1.0.0 and 1.0.1, the
strings "1" and "1.0" will load 1.0.1, while "1.0.0" will load 1.0.0.

Omitting the version entirely will pull from the master branch, which
now should always point to the latest release.  A branch name can also be
used, e.g. "dev", which will pull from that branch.