Generalized harvestBirchTree to handle all "straight" trees.
"Straight" trees are trees that always growth with a radius of one log and
do not branch (e.g. birch and pine).  Oak and jungle are NOT "straight"
trees, as both have larger forms with complex branching.

Introduced local table that can be expanded with "straight" trees from
modpacks.  A sapling table maps a log with the appropriate sapling for
replanting.