Major overhaul of igfarm API.
Functions that were passing state around are now encapsulated by the
Harvest class.  The farmGeneric function instantiates the call when called,
effectively storing the function arguments to be accessed by all of the
instance methods.  The change to the API is backward-incompatible
(arguments are passed as a table for naming, as there were too many
to be reasonably used as positional arguments).

The arguments to farmGeneric are now also passed to the callback, allowing
callback-specific (named) arguments to be passed through.  This way, for
example, a custom table of mod-specific logs and saplings can be passed to
the harvestStraightTree() callback function, while a custom table of
mod-specific seeds and crops can be passed to the (eventual rewrite of the)
_farmPlant() callback function.

The farm and harvest wrapper scripts have been updated accordingly.