Fix improper minimum fuel calculation for farm algorithm w.r.t. trees.
During the generic farming algorithm, before each movement the turtle
calculates the minimum necessary fuel to progess and compares it to its
current fuel.  This is simply the taxi-cab distance to the refueling
location plus a minfuel parameter.  If the minfuel parameter is unset, a
default of zero is used.  If the farming program is a tree program, this
fails to take into account the action of harvesting a tree, which costs the
height of the tree (less the first two blocks) times 2 (for moving up and
down).

Since modded trees could grow to arbitrarily up to the height limit,
having a minimum preset minfuel parameter is not practical.  Instead, the
Harvest class is changed to allow the farming callback to dynamically
change the minfuel level.  Every time the harvestStraightTree callback
harvests a tree, it compares (twice) the tree's height to the minfuel
level and updates the parameter as needed.

This won't accomodate tall trees early on, but over time the turtle will
learn the tallest a tree can get, and it will more accurately estimate its
necessary remaining fuel.