Huge restructuring of base API for unit testing.
The base ig module is now structured to be importable by a standard Lua
interpreter.  Unit tests for the base module (built with the lua-mock
public domain module, which lives in a subdirectory of test) live in the
test directory.

All functions are declared local and only exposed globally as part of the
public API section.  If loaded in a ComputerCraft context, "global"
variables are declared with the public function names.  If loaded in the
standard Lua context, a table with the member variables matching the public
function names is returned.

Indentation has been uniformly updated to 4 spaces.